move some util functions
